moving cumprod outside of affine
