+++ cvxpy/atoms/cumprod.py
@@ -13,16 +13,12 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 """
-from typing import List, Tuple
+from typing import Tuple
 
 import numpy as np
 
-import cvxpy.lin_ops.lin_op as lo
-import cvxpy.lin_ops.lin_utils as lu
 from cvxpy.atoms.affine.affine_atom import AffAtom
 from cvxpy.atoms.axis_atom import AxisAtom
-from cvxpy.atoms.gmatmul import gmatmul
-from cvxpy.constraints.constraint import Constraint
 from cvxpy.expressions.expression import Expression
 
 
@@ -41,7 +37,7 @@ class cumprod(AffAtom, AxisAtom):
         super(cumprod, self).__init__(expr, axis)
 
     @AffAtom.numpy_numeric
-    def numeric(self, values):
+    def numeric(self, values) -> np.ndarray:
         """
         Returns the cumulative product of the elements of the expression.
         """
@@ -51,7 +47,7 @@ class cumprod(AffAtom, AxisAtom):
         """The same as the input."""
         return self.args[0].shape
 
-    def _grad(self, values):
+    def _grad(self, values) -> list:
         """Gives the (sub/super)gradient of the atom w.r.t. each argument.
 
         Matrix expressions are vectorized, so the gradient is a matrix.
@@ -65,30 +61,6 @@ class cumprod(AffAtom, AxisAtom):
         # TODO implement grad
         return []
 
-    def get_data(self):
+    def get_data(self) -> list:
         """Returns the axis being summed."""
         return [self.axis]
-
-    def graph_implementation(
-        self, arg_objs, shape: Tuple[int, ...], data=None
-    ) -> Tuple[lo.LinOp, List[Constraint]]:
-        """
-        Cumulative product via difference matrix.
-
-        Parameters
-        ----------
-        arg_objs : list
-            LinExpr for each argument.
-        shape : tuple
-            The shape of the resulting expression.
-        data :
-            Additional data required by the atom.
-
-        Returns
-        -------
-        tuple
-            (LinOp for objective, list of constraints)
-        """
-        n = arg_objs[0]
-        A = lu.create_const(np.triu(np.ones(shape)), shape=shape)
-        return (gmatmul(A, n), [])
