+++ cvxpy/reductions/dgp2dcp/canonicalizers/cumprod_canon.py
@@ -1,0 +1,20 @@
+"""
+Copyright 2024 the CVXPY developers
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+"""
+def cumprod_canon(expr, args):
+    out = 0.0
+    for x_i, p_i in zip(args[0], expr.p):
+        out += p_i * x_i
+    return (1 / sum(expr.p)) * out, []
